Easier javascript shit than that horror show of typescript
