+++ samples/BlazorServerSide/Pages/_Host.cshtml
@@ -26,9 +26,11 @@
 
     <script src="_framework/blazor.server.js"></script>
 
-    <!-- Penman.Blazor.Quill scripts -->
-    <script src="_content/Penman.Blazor.Quill/quill-blot-formatter.min.js"></script>
+<!-- Penman.Blazor.Quill scripts -->
+    <script src="https://cdn.quilljs.com/1.3.6/quill.js"></script>
     <script src="_content/Penman.Blazor.Quill/penman-blazor-quill.js"></script>
+    <script src="_content/Penman.Blazor.Quill/quill-blot-formatter.min.js"></script>
+
 
 </body>
 </html>
