--- src/Penman.Blazor.Quill/jslib/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "quill-editor-imageupload",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "build": "webpack --config webpack.config.js",
-    "start": "webpack serve --config webpack.config.js --open"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.18.9",
-    "@babel/preset-env": "^7.18.9",
-    "@babel/preset-typescript": "^7.18.6",
-    "@types/lodash": "^4.14.182",
-    "@types/quill": "^2.0.9",
-    "babel-loader": "^8.2.5",
-    "copy-webpack-plugin": "^9.0.1",
-    "css-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.3.2",
-    "source-map-loader": "^4.0.0",
-    "style-loader": "^3.2.1",
-    "terser-webpack-plugin": "^5.3.3",
-    "ts-loader": "^9.3.1",
-    "typescript": "^4.7.4",
-    "url-loader": "^4.1.1",
-    "webpack": "^5.73.0",
-    "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^4.3.1"
-  },
-  "dependencies": {
-    "lodash": "^4.17.21",
-    "parchment": "^2.0.1",
-    "quill": "^1.3.7",
-    "quill-blot-formatter": "^1.0.5"
-  },
-  "keywords": [
-    "typescript",
-    "QuillJS",
-    "webpack",
-    "example"
-  ],
-  "description": "A basic typescript app with Quill additional functionality built on top of it",
-  "homepage": "https://github.com/domingoladron/quill-editor-imageupload",
-  "license": "MIT",
-  "author": {
-    "name": "Chris Laine (domingoladron)",
-    "url": "https://christopherlaine.net"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/domingoladron/quill-editor-imageupload.git"
-  }
-}
