--- src/Penman.Blazor.Quill/jslib/src/css/main.css
@@ -1,7 +1,0 @@
-html, body {
-	width: 100%;
-	height: 100%;
-	margin: 0;
-	padding: 0;
-	overflow: hidden;
-}
