--- src/Penman.Blazor.Quill/jslib/src/index.html
@@ -1,60 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <link href="https://cdn.quilljs.com/1.3.6/quill.snow.css" rel="stylesheet">
-  <link href="https://cdn.quilljs.com/1.3.6/quill.bubble.css" rel="stylesheet">
-</head>
-<body>
-<script>
-    window.onload = function(){
-        window.QuillFunctions.createQuill(
-            '#editor-container',
-            '#toolbar',
-            false,
-            'test',
-            'snow',
-            'debug',
-            '#editor-wrapper',
-            true,
-            '/images/',
-            '',
-            null);
-    };
-
-</script>
-<div id="editor-wrapper">
-  <div id="editor-container">
-    <div id="toolbar">
-      <select class="ql-header">
-        <option selected=""></option>
-        <option value="1"></option>
-        <option value="2"></option>
-        <option value="3"></option>
-        <option value="4"></option>
-        <option value="5"></option>
-      </select>
-
-      <span class="ql-formats">
-                  <button class="ql-bold"></button>
-                  <button class="ql-italic"></button>
-                  <button class="ql-underline"></button>
-                  <button class="ql-strike"></button>
-              </span>
-      <span class="ql-formats">
-                  <select class="ql-color"></select>
-                  <select class="ql-background"></select>
-              </span>
-      <span class="ql-formats">
-                  <button class="ql-list" value="ordered"></button>
-                  <button class="ql-list" value="bullet"></button>
-              </span>
-      <span class="ql-formats">
-                  <button class="ql-link"></button>
-                  <button class="ql-image"></button>
-      </span>
-    </div>
-  </div>
-</div>
-</body>
-</html>
