--- src/Penman.Blazor.Quill/wwwroot/penman-blazor-quill.js.LICENSE.txt
@@ -1,6 +1,0 @@
-/*!
- * Quill Editor v1.3.7
- * https://quilljs.com/
- * Copyright (c) 2014, Jason Chen
- * Copyright (c) 2013, salesforce.com
- */
